+++ .eslintrc.json
@@ -1,6 +1,37 @@
 {
   "extends": "next/core-web-vitals",
+  "plugins": [
+    "import"
+  ],
   "rules": {
-    "react/no-unescaped-entities": "off"
+    "react/no-unescaped-entities": "off",
+    "import/order": [
+      "error",
+      {
+        "groups": [
+          "builtin",
+          "external",
+          "internal",
+          [
+            "parent",
+            "sibling",
+            "index"
+          ],
+          "object",
+          "type"
+        ],
+        "pathGroups": [
+          {
+            "pattern": "react",
+            "group": "external",
+            "position": "before"
+          },
+          {
+            "pattern": "^components/(.*)$",
+            "group": "internal"
+          }
+        ]
+      }
+    ]
   }
 }
