+++ package.json
@@ -22,6 +22,7 @@
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
+    "eslint-plugin-import": "^2.29.1",
     "sass": "^1.77.0",
     "typescript": "^5"
   }
