--- src/app/_data/socialMediaInfo.json
@@ -1,22 +1,0 @@
-[
-  {
-    "name": "twitter",
-    "handle": "@biliscas",
-    "url": "https://x.com/biliscas"
-  },
-  {
-    "name": "linkedin",
-    "handle": "in/lucas-migliori",
-    "url": "https://www.linkedin.com/in/lucas-migliori/"
-  },
-  {
-    "name": "instagram",
-    "handle": "@lucas.migliori",
-    "url": "https://www.instagram.com/lucas.migliori/"
-  },
-  {
-    "name": "youtube",
-    "handle": "@lucasmigliori",
-    "url": "https://www.youtube.com/@lucasmigliori"
-  }
-]
