+++ src/app/_data/socialMediaInfo.ts
@@ -1,0 +1,22 @@
+export const socialMediaInfo = [
+  {
+    name: "twitter",
+    handle: "@biliscas",
+    url: "https://x.com/biliscas",
+  },
+  {
+    name: "linkedin",
+    handle: "in/lucas-migliori",
+    url: "https://www.linkedin.com/in/lucas-migliori/",
+  },
+  {
+    name: "instagram",
+    handle: "@lucas.migliori",
+    url: "https://www.instagram.com/lucas.migliori/",
+  },
+  {
+    name: "youtube",
+    handle: "@lucasmigliori",
+    url: "https://www.youtube.com/@lucasmigliori",
+  },
+];
