+++ src/app/lib/constants.ts
@@ -4,59 +4,3 @@ export const navLinks = [
   { href: "/projects", label: "Projects" },
   { href: "/contact", label: "Contact" },
 ];
-
-export const socialMediaInfo = [
-  {
-    name: "twitter",
-    handle: "@biliscas",
-    url: "https://x.com/biliscas",
-  },
-  {
-    name: "linkedin",
-    handle: "in/lucas-migliori",
-    url: "https://www.linkedin.com/in/lucas-migliori/",
-  },
-  {
-    name: "instagram",
-    handle: "@lucas.migliori",
-    url: "https://www.instagram.com/lucas.migliori/",
-  },
-  {
-    name: "youtube",
-    handle: "@lucasmigliori",
-    url: "https://www.youtube.com/@lucasmigliori",
-  },
-];
-
-export const mySkills = [
-  {
-    icon: "frontend",
-    title: "Frontend Mastery",
-    description:
-      "Expertise in TypeScript, React, Next.js, SASS, and the core web technologies (HTML, CSS, JS)",
-  },
-  {
-    icon: "stack",
-    title: "Stack Proficiency",
-    description:
-      "Familiarity with Node.js, Express.js, NestJS, SQL Server, ORMs, MongoDB and so on.",
-  },
-  {
-    icon: "aws",
-    title: "Cloud Knowledge",
-    description:
-      "AWS Certified Cloud Practitioner, with hands-on experience in EC2, S3, SES, and Lambda.",
-  },
-  {
-    icon: "git",
-    title: "Version Control",
-    description:
-      "Able to use structured project patterns like Atomic Design and GitFlow for robust development.",
-  },
-  {
-    icon: "test",
-    title: "QA Excellence",
-    description:
-      "Skilled in test planning, API testing, Postman scripting, and test automation with Cypress.",
-  },
-];
